Refactor circle playback into Player
